+++ doc/CHANGES.rst
@@ -19,6 +19,7 @@ Futur
 * [ADD]: add ``dramatiq-broker`` in configuration groups of pyramid and gunicorn
   apps
 * [DEL] Remove configuration group definition ``preload``
+* [REF] use ``configuration_post_load`` function to initialize services
 
 0.7.2 (2017-10-18)
 ------------------